pyproject.toml
[tool.poetry]
name = "flood-detection"
version = "0.1.0"
description = "Classify flood related text, extract locations, and visualize the results"
authors = ["Yasser Kaddoura <yasser.kaddoura19@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "flood_detection"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
tweepy = "^4.11.0"
python-dotenv = "^0.21.0"
numpy = "^1.23.4"
pandas = "^1.5.1"
matplotlib = "^3.6.1"
deep-translator = "^1.9.0"
transformers = "^4.23.1"
sklearn = "^0.0"
boto3 = "^1.25.2"
spacy = {version = "^3.4.3"}
dvc = "^2.31.0"
hydra-core = "^1.2.0"
click = "^8.1.3"
tqdm = "^4.64.1"
geopy = "^2.2.0"
plotly = "^5.10.0"
requests = "^2.28.1"
dash = "^2.6.2"
gensim = "^4.2.0"
dvc-gdrive = "^2.19.0"
langdetect = "^1.0.9"
sagemaker = "^2.116.0"
datasets = {version = "2.5.2", extras = ["s3"]}
torch = "^1.13.0"
imblearn = "^0.0"
dash-leaflet = "^0.1.23"
dash-extensions = "^0.1.6"
geobuf = "^1.1.1"
nltk = "^3.7"
dash-bootstrap-components = "^1.2.1"
sentence-transformers = "^2.2.2"
rake-nltk = "^1.0.6"
pillow = "^9.3.0"
cryptography = "^38.0.3"
geopandas = "^0.12.1"
rtree = "^1.0.1"
keybert = {extras = ["gensim"], version = "^0.7.0"}


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pandas-stubs = "^1.5.1.221024"
pre-commit = "^2.20.0"
flake8 = "^5.0.4"
setuptools = "^65.5.0"
visidata = "^2.10.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
